# backend/services/userService/models/user.go
package models

import (
	"time"
)

type User struct {
	Id        int        `json:"id" example:"1"`
	Email     string     `json:"email" binding:"required" example:"testuser"`
	Firstname string     `json:"firstname" binding:"required" example:"testuser"`
	Lastname  string     `json:"lastname" binding:"required" example:"testuser"`
	Password  string     `json:"-"`
	CreatedAt *time.Time `json:"created_at,string,omitempty" example:"2023-12-01T12:37:59.008583Z"`
	UpdatedAt *time.Time `json:"updated_at_at,string,omitempty" example:"2023-12-01T12:37:59.008583Z"`
}

type RegisterUserInput struct {
	Firstname string `json:"firstname" binding:"required" example:"testuser"`
	Lastname  string `json:"lastname" binding:"required" example:"testuser"`
	Email     string `json:"email" binding:"required" example:"testuser@test.de"`
	Password  string `json:"password" binding:"required" example:"test"`
}

type SignInUserInput struct {
	Email    string `json:"email" binding:"required" example:"testuser@test.de"`
	Password string `json:"password" binding:"required" example:"test"`
}

type RegisterUserOutput struct {
	UserId int `json:"userId" binding:"required" example:"1"`
}